bridge/requirements.txt
pystray
Pillow
pyodbc
firebase-admin
psutil
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
